+++ saritasa_invocations/pre_commit.py
@@ -2,7 +2,13 @@ import os
 
 import invoke
 
-from . import printing
+from . import _config, printing
+
+
+def run_pre_commit_cmd(context: invoke.Context, cmd: str, **kwargs) -> None:
+    """Run a pre-commit command."""
+    config = _config.Config.from_context(context)
+    context.run(command=f"{config.pre_commit.entry} {cmd}", **kwargs)
 
 
 @invoke.task
@@ -15,7 +21,7 @@ def install(
     hooks_str = (
         " ".join(f"--hook-type {hook}" for hook in hooks) if hooks else ""
     )
-    context.run(f"pre-commit install {hooks_str}")
+    run_pre_commit_cmd(context=context, cmd=f"install {hooks_str}")
 
 
 @invoke.task
@@ -28,20 +34,26 @@ def uninstall(
     hooks_str = (
         " ".join(f"--hook-type {hook}" for hook in hooks) if hooks else ""
     )
-    context.run(f"pre-commit uninstall {hooks_str}")
+    run_pre_commit_cmd(context=context, cmd=f"uninstall {hooks_str}")
 
 
 @invoke.task
 def run_hooks(
     context: invoke.Context,
-    hook_stage: str = "push",
+    hook_stage: str = "",
     params: str = "",
     skip: str = "",
 ) -> None:
     """Run all hooks against all files."""
     printing.print_success("Running pre-commit hooks")
-    context.run(
-        f"pre-commit run --hook-stage {hook_stage} --all-files {params}",
+    config = _config.Config.from_context(context)
+
+    if not hook_stage:
+        hook_stage = config.pre_commit.default_hook_stage
+
+    run_pre_commit_cmd(
+        context=context,
+        cmd=f"run --hook-stage {hook_stage} --all-files {params}",
         env={
             "SKIP": skip or os.environ.get("SKIP", ""),
         },
@@ -52,4 +64,4 @@ def run_hooks(
 def update(context: invoke.Context) -> None:
     """Update pre-commit dependencies."""
     printing.print_success("Updating pre-commit")
-    context.run("pre-commit autoupdate")
+    run_pre_commit_cmd(context=context, cmd="autoupdate")